#42027 Make Redactions scale with video resize
